\name{ArrayTV-package}
\alias{ArrayTV-package}
\alias{ArrayTV}
\docType{package}
\title{
ArrayTV-implementation of GC correction for arrays
}
\description{
Correct gc biases in array data. The user must provide a range of windows and this function will compute the tv score for each window and correct the arrays based upon the window with the highest tv score (i.e with window showing the most gc bias)
}
\details{
\tabular{ll}{
Package: \tab ArrayTV\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-08-02\cr
License: \tab Johsn Hopkins School of Medicine\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
Eitan Halper-Stromberg
Maintainer: Eitan Halper-Stromberg
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
type ?gcCorrectMain.Rd at the command prompt for an example
}
