\name{gcCorrect-methods}
\docType{methods}
\alias{gcCorrect-methods}
\alias{gcCorrect,matrix-method}
\title{
gcCorrect
}
\description{
methods to call gcCorrectMain for array data with varying formats
}
\section{Methods}{
\describe{

\item{\code{signature(object = "matrix")}}{
the array data
}
}}
\keyword{methods}
